Dividing logic for each component
